Fix database relation state handling (#75)
## Issue
Fixes #73, fixes #78

Database relations can be in many states:
- Relation missing
- Relation created but no data in databag or some required keys missing
from databag (#78)—we need to wait for related application
- Relation data valid from database_provides remote application but we
haven't created user yet—we need to create database/user & populate
databag
- Relation data valid from database_provides remote application and
we've created a user—do nothing
- Relation breaking—remote databag not accessible
(https://bugs.launchpad.net/juju/+bug/1960934, #73); we need to clean up
user if we created one

Before this PR, the `_Relation` class could be in multiple states and
some properties could be accessed but some properties would cause an
exception if accessed depending on the state. In other words, which
properties were available was implicit.

For example, these properties

https://github.com/canonical/mysql-router-k8s-operator/blob/92b029efd1bc0c183328e98f7fdd43cbd32da4b1/src/relations/database_requires.py#L44-L62
can only be accessed if

https://github.com/canonical/mysql-router-k8s-operator/blob/92b029efd1bc0c183328e98f7fdd43cbd32da4b1/src/relations/database_requires.py#L64-L66
is False and these conditions are met

https://github.com/canonical/mysql-router-k8s-operator/blob/92b029efd1bc0c183328e98f7fdd43cbd32da4b1/src/relations/database_requires.py#L102-L108

## Solution
Create different `_Relation` classes for different states so that all
properties/methods on the class can be called
e.g.

https://github.com/canonical/mysql-router-k8s-operator/blob/5a1757ba5d532e8bb4896972a6527bb0cb4e674d/src/relations/database_provides.py#L57

https://github.com/canonical/mysql-router-k8s-operator/blob/5a1757ba5d532e8bb4896972a6527bb0cb4e674d/src/relations/database_provides.py#L103

Also raise exceptions during the `__init__` of those objects so that we
don't have to check for exceptional cases (e.g. required key missing
from relation databag, relation is breaking) when using the object—if
the object initializes successfully, all of its properties/methods are
safe to use