Fix health check race condition
If the mysql <-> mysqlrouter relation is completed after the execution of the previous event and during the update status, start, or leader elected event, mysqlrouter will not be part of the cluster and the health check will fail.

Instead of running set_status during update status, start, or leader elected, run the entire reconcile handler to avoid this race condition.