Recover from hook errors when creating/deleting MySQL users (#165)
Fixes #157

Previously, the charm assumed that if a user was created or deleted that
the databag would be updated accordingly. However, if a user is created
or deleted and the event/hook fails, the user creation/deletion will go
through for MySQL but the databag will not be updated (since the event
failed). When the event is retried, the charm tried to create a user
that already exists or delete a user that doesn't exist