Bootstrap with force (#187)
Instead of relying on the existence of old router user to cleanup router
from cluster metadata, force bootstrap so that the metadata does not
need to be cleaned up.

Fixes https://github.com/canonical/mysql-k8s-operator/issues/345. The
issue was that the router charm would delete the user & router metadata,
but that only the user deletion would go through. Then, on the router
charm's next hook, the bootstrap failed because the metadata was not
cleaned up.

Additional context:
https://chat.canonical.com/canonical/pl/temrphcp3in5xqftkaxgowqtkr