Retry if MySQL Server is unreachable (#190)
Fixes #86

## Context
These are the situations the charm should handle without raising an
uncaught exception:
- `mysqlrouter --bootstrap` fails because server does not have quorum
- `mysqlrouter --bootstrap` fails because server is unreachable (error
code
[2003](https://dev.mysql.com/doc/mysql-errors/8.0/en/client-error-reference.html#error_cr_conn_host_error))
- mysqlsh `shell.connect()` fails because server is unreachable (error
code
[2003](https://dev.mysql.com/doc/mysql-errors/8.0/en/client-error-reference.html#error_cr_conn_host_error))

(mysqlsh commands seem to not fail even if server does not have quorum)

## Solution
Catch exceptions in the aforementioned situations & set waiting status.
Retry connections on the next Juju event

## Alternatives considered
We decided not to retry within the Juju event so that
- (primary reason) the user is not blocked from removing the unit(s)
- if the endpoint from mysql server changes, we need to get a new event
to get the updated endpoint

A retry within the Juju event with a short timeout was considered.
However, during initial deployment ~15 events are currently fired (so a
1 min timeout could block the charm for 15 minutes). It would be
possible to mitigate that by storing information about the last
retry—but the complexity was not worth it.

Discussion about Juju limitation that we may not get another event until
update status:
https://matrix.to/#/!xdClnUGkurzjxqiQcN:ubuntu.com/$_SXnGJWUQCeFGu9vpBxD5GJ5hgMMAHchOmWBPiqjDFY?via=ubuntu.com&via=matrix.org&via=cutefunny.art

## Implementation/refactoring
- Catch mysql shell `DBError` exceptions, serialize them to JSON, and
raise corresponding exception in charm code. This was done to move
complexity (i.e. exception handling) from mysqlsh python code to charm
code
- Move mysqlsh python code from strings to Jinja templates (for better
syntax highlighting/analysis & code review/diffs)
- Use JSON files to transfer data from mysqlsh python code to charm code
instead of printing & parsing stdout