[DPE-3771] Add node port support (#211)
DPE-3771: In case mysql-router is connected with any app that sets
`external-node-connectivity=true`, then NodePort must be configured and
shared via relation. Applications that do not set it will receive the
internal k8s DNS record instead. This way, an application, if running
within the same k8s cluster, can choose to go through the internal
cluster IPs and reduce its own latency.

An app outside of Juju ecosystem can connect with the database by
relating the router with data-integrator, which triggers the creation of
the NodePort. Then, the NodePort can be used to connect.

TLS is managed as well, in this change. If TLS and NodePort are both
enabled, then the workers' IPs and hostnames where the routers are
placed are captured and added as SANs to the certificate.

---------

Co-authored-by: Carl Csaposs <carl.csaposs@canonical.com>