[DPE-3899, DPE-4173] Try to clear connection pool before relating with COS to avoid TIME_WAIT connections; stabilize exporter tests (#245)
## Issue
https://github.com/canonical/mysql-router-k8s-operator/issues/219
https://github.com/canonical/mysql-router-k8s-operator/issues/229

We are running into a connection in `TIME_WAIT` when starting
`mysql_router_exporter`. This happens in both the K8s and VM charms. The
guess is that this is related to the connection pool we use in our
tests.

## Solution
Clear the connection pool before relating with COS, re-run exporter
tests 5-10 times, observe if tests have stabilized