feature: DPE-7534 use absolute fqdn and bigger ttl (#445)
## Issue

See issues #443 and #444

Also, mysql-router configure `ttl=5` [seconds] whenever clusterset is
configured, independently of the usage of `gr_notifications`.
For comparison, without `gr_notifications` default is `ttl=0.5`, so this
is a conservative change giving geographical concerns when having
multiple clusters.
Given we are using `gr_notifications`, it's safe to override to a
greater value, since this is a fallback mechanism.

## Solution

- use and propagate absolute fqdn (trailing dot) internally and to
consumers
- set `60s` as default metadata_cache refresh interval

For test bundle (3 router, 3 server, 1 test-app), this changes (+ usage
of absolute fqdn on server) took coredns requests from 2.18 kp/s to
0.270kp/s.


![requests_k8s](https://github.com/user-attachments/assets/db1d27bd-8b68-4a77-99b4-94fcb4446071)

Fixes #443 #444